+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 =========================================================
+2015-09-24 v0.4.12 release
+=========================================================
+
+Improve local-only cluster settings for KovertVertx and hazelcast (use tcp instead of multicast, set to loopback)
+Local only clustering is false by default
+
+=========================================================
 2015-09-24 v0.4.11 release
 =========================================================
 
-Add configuration option for KovertVerticle that can force networking to localhost for clustering (mac osx might still violate this and broadcast all interfaces)
-Add loading of default hazelcast cluster xml from vertx when using KovertVerticle and clustering mode
+Add configuration option for KovertVertx that can force networking to localhost for clustering (mac osx might still violate this and broadcast all interfaces)
+Add loading of default hazelcast cluster xml from vertx when using KovertVertx and clustering mode
 Sample App now has more realistic logging configuration, separate app.log, error.log and access.log
 
 =========================================================